Update to surrogates LDAP validation to search across all configured sources and fix to compile and tests
