+++ support/cas-server-support-surrogate-authentication-ldap/build.gradle
@@ -24,6 +24,7 @@ dependencies {
     implementation project(":core:cas-server-core-services-api")
 
     testImplementation project(":core:cas-server-core-authentication")
+    testImplementation project(":core:cas-server-core-services-authentication")
     testImplementation project(":core:cas-server-core-tickets")
     testImplementation project(":core:cas-server-core-notifications")
     testImplementation project(":core:cas-server-core-util")
